API refactoring Assets, Search, Users and Rules
