+++ app/urls.py
@@ -5,14 +5,18 @@ from django.conf.urls import include, url
 from django.contrib import admin
 from django.contrib.staticfiles.urls import staticfiles_urlpatterns
 from django.contrib.auth import views as auth_views
-# from django.contrib.auth import logout as v_logout
+from rest_framework_swagger.views import get_swagger_view
 from users import views as user_views
 
 
 def i18n_javascript(request):
     return admin.site.i18n_javascript(request)
 
+
+api_schema_view = get_swagger_view(title='PatrOwl Manager REST-API')
+
 urlpatterns = [
+    url(r'^apis-doc', api_schema_view),
     url(r'^admin/', admin.site.urls),
     url(r'^engines/', include('engines.urls')),
     url(r'^findings/', include('findings.urls')),
@@ -28,13 +32,9 @@ urlpatterns = [
 
     url(r'^login$', user_views.login, name='login'),
     url(r'^logout$', auth_views.logout, {'next_page': '/login'}, name='logout'),
-    # url(r'^logout$', v_logout, {'next_page': '/login'}, name='logout'),
     url(r'^signup$', user_views.signup, name='signup'),
 
     url(r'^admin/jsi18n/', i18n_javascript),
-
-    #url(r'^api-auth/', include('rest_framework.urls')),
-    # url(r'^api-auth/', include('rest_framework.urls')),
 ]
 
 # debug toolbar & download file
