+++ assets/templates/add-asset-owner.html
@@ -36,7 +36,7 @@
 </div> -->
 
 <ul class="breadcrumb">
-  <li><a href="/assets/list">assets</a></li>
+  <li><a href="{% url 'list_assets_view' %}">assets</a></li>
   <li><a href="/assets/owners/list">owners</a></li>
   <li class="active">add new asset owner [<a href='/assets/owners/list'>back</a>]</li>
 </ul>
