+++ assets/templates/add-asset.html
@@ -2,7 +2,7 @@
 {% block content %}
 
 <ul class="breadcrumb">
-  <li><a href="/assets/list">assets</a></li>
+  <li><a href="{% url 'list_assets_view' %}">assets</a></li>
   <li class="active">add new asset manually [<a href='/assets/list'>back</a>]</li>
 </ul>
