+++ assets/templates/add-assets-bulk.html
@@ -2,13 +2,13 @@
 {% block content %}
 
 <ul class="breadcrumb">
-  <li><a href="/assets/list">assets</a></li>
+  <li><a href="{% url 'list_assets_view' %}">assets</a></li>
   <li class="active">add new assets in bulk [<a href='/assets/list'>back</a>]</li>
 </ul>
 
 <div class="container">
   <form action="/assets/bulkadd" method="post" class="form-horizontal" enctype="multipart/form-data">
-    <legend>Add assets</legend>
+    <legend>Add assets in bulk</legend>
       {% csrf_token %}
       {% for field in form %}
           {% if field.errors %}
