+++ assets/templates/list-asset-owners.html
@@ -2,7 +2,7 @@
 {% block content %}
 
 <ul class="breadcrumb">
-  <li><a href="/assets/list">assets</a></li>
+  <li><a href="{% url 'list_assets_view' %}">assets</a></li>
   <li><a href="/assets/owners/list">owners</a></li>
   <li class="active">list</li>
 </ul>
