+++ assets/templates/report-assetgroup-findings.html
@@ -2,7 +2,7 @@
   <head>
     {% load patrowl_tags %}
     <meta http-equiv="Content-Type" content="text/html; charset=utf-8"></meta>
-    <title>PatrOwl Scan Report</title>
+    <title>PatrOwl Asset Group Report</title>
     <style type="text/css" media="all">
     	UL.ulist {padding: 0 10px; line-height:25px; margin-bottom:0px; margin-top:0px;};
     	LI.list  {padding: 0 10px; line-height:25px; margin-bottom:0px; margin-top:0px; list-style: disc;}
