+++ requirements.txt
@@ -5,37 +5,45 @@ billiard==3.5.0.3
 celery==4.0.2
 certifi==2017.4.17
 chardet==3.0.4
-Django==1.11.16
+coreapi==2.3.3
+coreschema==0.0.4
+Django==1.11.11
 django-celery-beat==1.0.1
 django-celery-results==1.0.1
+django-cursor-pagination==0.1.3
 django-datetime-widget==0.9.3
 django-debug-toolbar==1.8
 django-extensions==1.8.1
+django-rest-swagger==2.2.0
 djangorestframework==3.7.7
 flower==0.9.2
 future==0.16.0
 futures==3.1.1
 gunicorn==19.9.0
 idna==2.5
+itypes==1.1.0
+Jinja2==2.10
 jsonfield==2.0.2
 kombu==4.1.0
 libmagic==1.0
+MarkupSafe==1.1.0
 meld3==1.0.2
 netaddr==0.7.19
+openapi-codec==1.3.2
 psycopg2-binary==2.7.5
 python-magic==0.4.15
-#python-magic-bin==0.4.14
+python-magic-bin==0.4.14
 python-memcached==1.59
 pytz==2018.3
-requests==2.20.0
+requests==2.18.1
+simplejson==3.16.0
 singledispatch==3.4.0.3
 six==1.10.0
 sqlparse==0.2.3
 supervisor==3.3.3
 thehive4py==1.4.3
-#tornado==4.5.1
-tornado==5.1.1
+tornado==4.5.1
+uritemplate==3.0.0
 urllib3==1.21.1
 vine==1.1.4
 Werkzeug==0.12.2
-django-cursor-pagination
