+++ search/views.py
@@ -7,12 +7,9 @@ from engines.models import Engine, EnginePolicy
 from scans.models import ScanDefinition
 from findings.models import Finding
 
-def search_view(request):
-    results = []
-    kw = request.GET.get('srch-term', None)
 
-    if not kw:
-        return render(request, 'search-results.html', {'results': results})
+def _search(kw):
+    results = []
 
     # search by asset value or description
     for asset in Asset.objects.filter(
@@ -56,4 +53,24 @@ def search_view(request):
         Q(title__icontains=kw) | Q(description__icontains=kw) | Q(type__icontains=kw) | Q(solution__icontains=kw) | Q(vuln_refs__icontains=kw) | Q(links__icontains=kw) | Q(tags__icontains=kw)):
         results.append({"type": "finding", "value": finding.title, "id": finding.id, "link": "/findings/details/"+str(finding.id)})
 
+    return results
+
+
+def search_view(request):
+    kw = request.GET.get('srch-term', None)
+
+    if not kw:
+        return render(request, 'search-results.html', {'results': []})
+
+    results = _search(kw)
+    return render(request, 'search-results.html', {'results': results, 'search_term': kw})
+
+
+def search_api(request):
+    kw = request.GET.get('srch-term', None)
+
+    if not kw:
+        return JsonResponse()
+
+    results = _search(kw)
     return render(request, 'search-results.html', {'results': results, 'search_term': kw})
